move config to src
